Milestone: Fix "indexing" bug
So-called "indexing" bug was seen when looking at conflicts between
mostly overlapping molecules, including overlapping a fragment with
itself.

It seems to be called because of comparing the value of United-Atom
charges (via `getPreviewCharge()`) to all atom charges (`charges[atomid]`).

This seems to be fixed now but needs to be tested extensively.